--- examples/config/mvc/four_legs_example.yaml
@@ -1,8 +1,0 @@
-scene:
-  ANIMATED_CHARACTERS:
-    - character_cfg: examples/characters/char6/char_cfg.yaml
-      motion_cfg: examples/config/motion/zombie.yaml
-      retarget_cfg: examples/config/retarget/four_legs.yaml
-controller:
-  MODE: video_render
-  OUTPUT_VIDEO_PATH: ./video.gif
