+++ examples/quadruped/four_legs_example.yaml
